requirements.txt
quart==0.20.0
hypercorn==0.16.0
quart-auth==0.9.0
quart-cors==0.7.0
quart-wtforms==1.0.3
alembic==1.15.2
SQLAlchemy[asyncio]==2.0.40
aiosqlite==0.20.0
asyncpg==0.31.0
aiomysql==0.3.0
psycopg2-binary==2.9.10
pymysql==1.1.1
greenlet==3.2.2
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.3.2
aiofiles==24.1.0
aiosmtplib==5.1.0
cryptography==44.0.3
PyJWT==2.10.1
itsdangerous==2.2.0
WTForms==3.2.1
email-validator==2.2.0
cloudinary==1.44.0
pysubs2==1.8.0
lxml==5.3.0
chardet==5.2.0
kitsu-py==1.1.2
themoviedb==1.0.2
pyMALv2==0.0.3
iso639-lang==2.6.0
python-dotenv==1.0.1
click==8.1.8
dacite==1.9.2
python-dateutil==2.9.0.post0
six==1.17.0
certifi==2025.4.26
idna==3.10
charset-normalizer==3.4.2
urllib3==2.6.3
attrs==25.3.0
frozenlist==1.6.0
multidict==6.4.3
propcache==0.3.1
yarl==1.20.0
typing_extensions==4.13.2
Jinja2==3.1.6
MarkupSafe==3.0.2
Mako==1.3.10
Werkzeug==3.1.5
Brotli==1.2.0
zstandard==0.23.0
rarfile==4.2
rapidfuzz==3.10.1
logtail-python==0.2.9
cachelib==0.9.0
resend==2.15.0